MapPosicion

+++ src/Fifa.Core/Posicion.cs
@@ -2,9 +2,9 @@ namespace Fifa.Core;
 
 public class Posicion
 {
-    public int IdPosicion { get; set; }
+    public byte IdPosicion { get; set; }
     public string Nombre { get; set; }
-    public Posicion(int idPosicion, string nombre)
+    public Posicion(byte idPosicion, string nombre)
     {
         IdPosicion = idPosicion;
         Nombre = nombre;
